Added some exception handling
