+++ Analogue-Clock.layout
@@ -1,40 +1,40 @@
 [Editor_0]
-CursorCol=3
-CursorRow=56
-TopLine=31
+CursorCol=16
+CursorRow=1
+TopLine=1
 LeftChar=1
 [Editors]
-Order=1,2,3,4,5,6,7,0
-Focused=-1
+Order=1,2,3,4,5,6,0
+Focused=3
 [Editor_1]
-CursorCol=15
-CursorRow=72
+CursorCol=16
+CursorRow=1
 TopLine=1
 LeftChar=1
 [Editor_2]
 CursorCol=3
 CursorRow=37
-TopLine=1
+TopLine=20
 LeftChar=1
 [Editor_3]
-CursorCol=20
-CursorRow=7
+CursorCol=1
+CursorRow=8
 TopLine=1
 LeftChar=1
 [Editor_4]
-CursorCol=19
+CursorCol=1
 CursorRow=10
 TopLine=1
 LeftChar=1
 [Editor_5]
-CursorCol=11
-CursorRow=16
+CursorCol=20
+CursorRow=1
 TopLine=1
 LeftChar=1
 [Editor_6]
-CursorCol=2
-CursorRow=26
-TopLine=13
+CursorCol=16
+CursorRow=61
+TopLine=97
 LeftChar=1
 [Editor_7]
 CursorCol=35
