+++ Source Files/clock.h
@@ -4,6 +4,9 @@
 
 #include <graphics.h>
 #include <time.h>
+#include "timeio.h"
+#include "structure.h"
+
 
 
 Point RotatePoint (Point Rotation, Point Center, double Radian)
@@ -137,22 +140,7 @@ void DigitalClock (Time T, int Period)
 }
 
 
-Time Initiate ()
-{
-	time_t t;
-    struct tm* CurrentTime;
 
-    t = time(NULL);
-    CurrentTime = localtime(&t);
-    
-    Time T;
-    
-    T.Hour = CurrentTime->tm_hour;
-    T.Minute = CurrentTime->tm_min;
-    T.Second = CurrentTime->tm_sec;
-  
-    return T;
-}
 
 
 Hand DrawHands (Point Center, Time T, int Period)
@@ -295,6 +283,104 @@ void LiveClock (Point Center, Time T, Hand ClockHands, int Formate, int *Period)
 		DrawHands(Center, T, *Period);
 		delay(1000);
 	}
+}
+
+
+void MainDriver()
+{
+    int gm = DETECT, gd;
+	
+	initgraph(&gm, &gd, (char *)"");
+	cleardevice();
+	
+	system("pause");
+	system("cls");
+	
+	int Formate = 1, Period = 0;
+	Time T=Initiate();
+	Point Center;
+	char Ans;
+	
+	Center.X = (getmaxx() / 2);
+	Center.Y = (getmaxy() / 2);
+	
+	Center.Y -= 75;
+	
+	DrawClock(Center, 1);
+	
+	if(Start())
+	{
+		printf("\n");
+		printf("\tGreat User ! Press Any Key Any Time To Pause The Clock !\n");
+		
+		LiveClock(Center, T, DrawHands (Center, T, Period), Formate, &Period);
+		
+		printf("\n");
+		printf("\tDo You Want To Continue With Custom Time? (Y/N) --> ");
+		scanf(" %c", &Ans);
+		
+		if(toupper(Ans) != 'Y')
+		{
+		    fflush(stdin);
+			printf("\n");
+		    
+		    printf("\tGood Bye User ! ");
+		    system("pause");
+		    
+			return;
+	    }
+	}
+	
+	closegraph();
+	fflush(stdin);
+	
+	initgraph(&gm, &gd, (char *)"");
+	cleardevice();
+	
+	DrawClock(Center, 2);
+	system("cls");
+	
+	printf("\n\t");
+	printf("!..... Analogue Clock..... !");
+	
+	printf("\n\n");	
+	
+	GetFormate (&Formate);
+	printf("\n");
+	
+	while(1)
+	{
+		if(Formate == 1)
+		{
+			system("cls");
+			
+			IfValidFormate (Formate);
+			GetTime(&T, Formate, &Period);
+			
+			break;
+		}
+		
+		if(Formate == 2)
+		{
+			system("cls");
+			
+			IfValidFormate (Formate);
+			GetTime(&T, Formate, &Period);
+			
+			break;
+		}
+		
+		system("cls");
+		IfInvalidFormate (&Formate);
+	}
+	
+	printf("\n");
+	printf("\tGreat User ! Press Any Key Any Time To Close The Program !\n");
+	
+	LiveClock(Center, T, DrawHands (Center, T, Period), Formate, &Period);
+	
+	getch();
+	closegraph();		
 }
